adapt 404, imprint page, remove old index page
