--- src/DEPRECATED_pages/404.js
@@ -1,14 +1,0 @@
-import React from "react"
-
-import Layout from "../DEPRECATED_components/layout"
-import SEO from "../DEPRECATED_components/seo"
-
-const NotFoundPage = () => (
-  <Layout>
-    <SEO title="404: Not found" />
-    <h1>NOT FOUND</h1>
-    <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </Layout>
-)
-
-export default NotFoundPage
