+++ src/index.story.js
@@ -1,6 +1,8 @@
 import React from "react"
 
+import Imprint from "./pages/imprint"
 import Index from "./pages"
+import NotFound from "./pages/404"
 import Register from "./pages/register.js"
 import Values  from "./pages/values";
 
@@ -9,6 +11,8 @@ export default {
   title: "Pages|main",
 }
 
+export const ImprintPage = () => <Imprint />
 export const IndexPage = () => <Index />
+export const NotFoundPage = () => <NotFound />
 export const RegisterPage = () => <Register />
 export const ValuesPage = () => <Values />
