+++ src/oldpages.story.js
@@ -1,8 +1,7 @@
 import React from "react"
 
-import NotFound from "./DEPRECATED_pages/404"
 import HelpPullRequest from "./DEPRECATED_pages/HelpPullRequest"
-import Imprint from "./DEPRECATED_pages/imprint"
+import TheTeam from "./DEPRECATED_pages/theteam"
 import PrivacyPolicy from "./DEPRECATED_pages/privacypolicy"
 import Register from "./DEPRECATED_pages/registration"
 import Sponsoring from "./DEPRECATED_pages/sponsoring"
@@ -12,9 +11,8 @@ export default {
   title: "Pages|old",
 }
 
-export const NotFoundPage = () => <NotFound />
 export const HelpPullRequestPage = () => <HelpPullRequest />
-export const ImprintPage = () => <Imprint />
+export const TheTeamPage = () => <TheTeam />
 export const PrivacyPolicyPage = () => <PrivacyPolicy />
 export const RegisterPage = () => <Register />
 export const SponsoringPage = () => <Sponsoring />
